+++ R/AbstractAnnData.R
@@ -271,18 +271,21 @@ AbstractAnnData <- R6::R6Class(
     #'
     #' @param file See [as_HDF5AnnData()]
     #' @param compression See [as_HDF5AnnData()]
+    #' @param chunk_size See [as_HDF5AnnData()]
     #' @param mode See [as_HDF5AnnData()]
     #'
     #' @return An [`HDF5AnnData`] object
     as_HDF5AnnData = function(
       file,
       compression = c("none", "gzip", "lzf"),
+      chunk_size = "auto",
       mode = c("w-", "r", "r+", "a", "w", "x")
     ) {
       as_HDF5AnnData(
         adata = self,
         file = file,
         compression = compression,
+        chunk_size = chunk_size,
         mode = mode
       )
     },
@@ -293,15 +296,23 @@ AbstractAnnData <- R6::R6Class(
     #'
     #' @param path See [write_h5ad()]
     #' @param compression See [write_h5ad()]
+    #' @param chunk_size See [write_h5ad()]
     #' @param mode See [write_h5ad()]
     #'
     #' @return `path` invisibly
     write_h5ad = function(
       path,
       compression = c("none", "gzip", "lzf"),
+      chunk_size = "auto",
       mode = c("w-", "r", "r+", "a", "w", "x")
     ) {
-      write_h5ad(object = self, path, compression = compression, mode = mode)
+      write_h5ad(
+        object = self,
+        path,
+        compression = compression,
+        chunk_size = chunk_size,
+        mode = mode
+      )
     }
   ),
   private = list(