+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fix CI (PR #418).
 - Add continuous benchmarking using bencher (PR #423, PR #425).
+- Allow manually setting chunk size for HDF5 writes (PR #424).
 - Add auto-chunking for HDF5 writes to improve performance (PR #424).
 
 # anndataR 1.1.0