+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fix CI (PR #418).
 - Add continuous benchmarking using bencher (PR #423, PR #425).
+- Handle unnamed `SingleCellExperiment` assays in `as_AnnData()` by automatically assigning names with a warning (PR #420).
 - Optimise sparse matrix reading performance by avoiding `Matrix::sparseMatrix` and constructing objects manually (PR #417).
 
 # anndataR 1.1.0